% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step_approx.R
\name{cbbinom.logl}
\alias{cbbinom.logl}
\title{Cost function for optimizing tau for spike-in data}
\usage{
cbbinom.logl(p, z, y, prob, c)
}
\description{
Cost function for optimizing tau for spike-in data
}
