% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.R
\name{dBBNB}
\alias{dBBNB}
\title{Probablity mass function of the distribution for observed data z_{ij}}
\usage{
dBBNB(z, pi0, mu, size, CE, rho, GQ.object, EY = FALSE)
}
\description{
Probablity mass function of the distribution for observed data z_{ij}
}
