% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decent.R
\name{decent}
\alias{decent}
\title{DECENT}
\usage{
decent(data.obs, cell.type, spikes = NULL, spike.conc = NULL,
  CE.range = c(0.02, 0.1), use.spikes = FALSE, normalize = "ML",
  GQ.approx = TRUE, maxit = 30, parallel = T, n.cores = 0,
  imputed = F, dir = "./")
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells.}

\item{cell.type}{A factor or a integer/numeric vector starting from 1 providing cell-type labels.}

\item{spike.conc}{A vector of theoretical count for each spike-in in one cell (ONLY if spikes = \code{TRUE}).}

\item{CE.range}{A two-element vector of the lower limit and upper limit for the estimated
capture efficiencies (ONLY if spikes = \code{FALSE}, default [0.02, 0.10]).}

\item{use.spikes}{If \code{TRUE}, use spike-ins to estimate capture efficiencies.}

\item{normalize}{Normalization method, either 'ML' (maximum likelihood) or 'TMM' (Robinson et al., 2010).}

\item{GQ.approx}{If \code{TRUE}, use GQ approximation to speed up E-step.}

\item{parallel}{If \code{TRUE}, run DECENT in parallel.}

\item{n.cores}{Number of CPU cores to use, default is all (ONLY if \code{TRUE}).}

\item{imputed}{If \code{TRUE}, include imputed data matrix in the output (Not available now).}

\item{dir}{Output directory for the DE model estimates, no-DE model estimates as well as the LRT output.}

\item{spike}{Observed count matrix for spike-ins, rows represent spike-ins, columns represent cells
(ONLY if spikes = \code{TRUE}).}
}
\value{
A data frame containing the result of differential expression analysis.
}
\description{
Differential Expression with Capture Efficiency adjustmeNT
}
