% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{EstepByGene}
\alias{EstepByGene}
\title{E-step without Gauss–Legendre quadrature approximation}
\usage{
EstepByGene(par, z, sf, pi0, mu, disp, k, b)
}
\arguments{
\item{par}{A matrix with two columns containing coefficient for the dropout model. The second column is currently unused
and set to zero but may be used in the future to model droput rates dependence on gene-specific factors.}

\item{z}{vector of observed count}

\item{sf}{Size factors for different cells}

\item{pi0}{Gene-specific zero-inflated parameter}

\item{mu}{Gene-specific mean parameter}

\item{disp}{Reciprocal of gene-specific size parameter for}
}
\description{
Perform E-step for one gene, without GQ approximation
}
