% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputed.R
\name{getImputed}
\alias{getImputed}
\title{Get imputed data}
\usage{
getImputed(data.obs, p, sf, CE, cell.type, parallel = T)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{p}{LRT estimates, using par.DE by default}

\item{sf}{A vector containing size factors for all cells}

\item{CE}{A  vector containing capture efficiencies for all cells}

\item{cell.type}{A factor or a integer/numeric vector starting from 1 giving cell-type labels}

\item{parallel}{If \code{TRUE}, run in parallel}
}
\value{
A matrix containing the imputed data assuming DE.
}
\description{
Get imputed data by running E-step once without GQ approximation using the LRT estimates.
}
