% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputed.R
\name{getSImputed}
\alias{getSImputed}
\title{Get imputed data}
\usage{
getSImputed(data.obs, par.est, sf, CE, XW, tau, parallel = T)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{par.est}{LRT/EM estimates, using par.DE by default}

\item{sf}{A vector containing size factors for all cells}

\item{CE}{A  vector containing capture efficiencies for all cells}

\item{XW}{A matrix containing information for X (cell-type) and W (other covariates)}

\item{tau}{cell-specific estimates (intcp,slope) that link Beta-Binomial dispersion parameter to the mean expression.}

\item{parallel}{If \code{TRUE}, run in parallel}

\item{M}{number of imputed values to be generated.}
}
\value{
A matrix containing the single imputed data.
}
\description{
Get single imputed data by sampling from the posterior distribution of the complete data given the observed data and LRT estimates.
}
