% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.R
\name{lrTest}
\alias{lrTest}
\title{Likelihood-ratio test}
\usage{
lrTest(data.obs, out, X, W = NULL, tau, parallel)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{out}{Output of fitNoDE, it contains EM algorithm estimates for models without DE between cell-types}

\item{X}{Data matrix containing dummy variables for covariates of interest (cell-type)}

\item{W}{Data matrix containing other covariates to adjust DE analysis. Default NULL}

\item{tau}{cell-specific estimates (intcp,slope) that link Beta-Binomial dispersion parameter to the mean expression.}

\item{parallel}{If \code{TRUE}, run in parallel}
}
\value{
A list containing statistics, p-values and parameter estimates for models without DE.
}
\description{
Likelihood-ratio test for DE analysis of scRNA-seq data
}
