% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.R
\name{SImputeByGene}
\alias{SImputeByGene}
\title{Perform single imputation for one gene using the fitted model}
\usage{
SImputeByGene(par, z, pi0, mu, sf, disp, k, b, M = 1)
}
\description{
Perform single imputation for one gene using the fitted model
}
