% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step_approx.R
\name{update.rho}
\alias{update.rho}
\title{Cost function for optimizing tau using endogenous genes}
\usage{
\method{update}{rho}(p, x, size, CE, sf = 1)
}
\description{
Cost function for optimizing tau using endogenous genes
}
