% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step_approx.R
\name{update.rho3}
\alias{update.rho3}
\title{Cost function for optimizing tau using endogenous genes}
\usage{
\method{update}{rho3}(p, z, size, CE)
}
\description{
Cost function for optimizing tau using endogenous genes
}
