% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{zinbGrad}
\alias{zinbGrad}
\title{Gradient of complete data log-likelihood function for DECENT model}
\usage{
zinbGrad(p, y, status, sf, ct)
}
\arguments{
\item{p}{A vector of parameter
p[1] = logistic regression intercept for pi0
p[2] = log mu for the first cell type
p[3] = log fold-change (FC) for mean parameter
p[4] = log dispersion (1/size) parameter}
}
\description{
Not used currently
}
