% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decent.R
\name{decent}
\alias{decent}
\title{DECENT}
\usage{
decent(data.obs, X, W = NULL, spikes = NULL, spike.conc = NULL,
  CE.range = c(0.02, 0.1), tau.init = c(-5, 0), tau.global = T,
  tau.est = "endo", use.spikes = FALSE, normalize = "ML",
  GQ.approx = TRUE, maxit = 30, parallel = T, n.cores = 0,
  s.imputed = F, E.imputed = F, dir = "./")
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells.}

\item{X}{An R model formula with covariates of interest (cell-type) as factor on the righthand side.}

\item{W}{An R model formula with other covariates to adjust DE analysis on the righthand side. Default NULL}

\item{spike.conc}{A vector of theoretical count for each spike-in in one cell (ONLY needed if spikes = \code{TRUE}).}

\item{CE.range}{A two-element vector of the lower limit and upper limit for the estimated range of
capture efficiencies (ONLY needed if spikes = \code{FALSE}, default [0.02, 0.10]).}

\item{tau.init}{initial estimates (intcp,slope) that link Beta-Binomial dispersion parameter to the mean expression.}

\item{tau.global}{whether to use the same tau parameters across cell. Default TRUE}

\item{tau.est}{Methods to estimate tau parameters. The default 'endo' corresponds to using endogeneous genes. Other options
are 'spikes' that corresponds to using spike-ins and 'none', which means tau.init is not further estimated.}

\item{use.spikes}{If \code{TRUE}, use spike-ins to estimate capture efficiencies.}

\item{normalize}{Method for estimating size factors, either 'ML' (maximum likelihood, Ye et al., 2017) or 'TMM' (Robinson et al., 2010).}

\item{GQ.approx}{If \code{TRUE}, use Gaussian-Quadrature approximation to speed up E-step.}

\item{parallel}{If \code{TRUE}, run DECENT in parallel.}

\item{n.cores}{Number of CPU cores to use, default is all (ONLY if parallel=\code{TRUE}).}

\item{s.imputed}{If \code{TRUE}, save the single imputed data matrix under the output diretory.}

\item{E.imputed}{If \code{TRUE}, save the mean imputed data matrix under the output diretory.}

\item{dir}{Directory to save all outputs, including EM algorithm estimates of no-DE model and the LRT output.}

\item{spike}{Observed count matrix for spike-ins, rows represent spike-ins, columns represent cells. Only needed if spikes = \code{TRUE}).}
}
\value{
A list containing the result of differential expression analysis, with the following components: stat,pval,par.DE and par.noDE.
}
\description{
Differential Expression with Capture Efficiency adjustmeNT
}
