% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step_approx.R
\name{Estep2ByGene}
\alias{Estep2ByGene}
\title{E-step with Gauss–Legendre quadrature approximation}
\usage{
Estep2ByGene(par, z, sf, pi0, mu, disp, k, b, GQ.object)
}
\arguments{
\item{par}{A matrix with two columns containing coefficient for the dropout model. The second column is currently unused
and set to zero but may be used in the future to model droput rates dependence on gene-specific factors.}

\item{z}{vector of observed count}

\item{sf}{Size factors for different cells}

\item{pi0}{gene-specific zero-inflated parameter}

\item{mu}{gene-specific mean parameter}

\item{disp}{Reciprocal of gene-specific size parameter for}

\item{GQ.object}{Gauss-Legendre quadrature points and weights}
}
\description{
Perform E-step for one gene, with GQ approximation
}
