% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{EstepByGene}
\alias{EstepByGene}
\title{E-step without Gauss–Legendre quadrature approximation}
\usage{
EstepByGene(par, z, z.ind, sf, pi0, mu, disp, y)
}
\arguments{
\item{par}{A vector of dropout model}

\item{z}{A matrix, 1st col = obs count, 2nd,3rd row = DO par}

\item{z.ind}{A binary indicator of (z == 0) in the original obs count (before imputation)}

\item{sf}{Size factors for different cells}

\item{mu}{Mean parameter for gene i}

\item{disp}{Reciprocal of size parameter for gene i}
}
\description{
Perform E-step for one gene, without GQ approximation
}
