% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_DE.R
\name{fitDE}
\alias{fitDE}
\title{Fitting DE model using EM algorithm}
\usage{
fitDE(data.obs, cell.type, spikes, spike.conc, CE.range, use.spikes, normalize,
  GQ.approx, maxit, parallel)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{cell.type}{A factor or a integer/numeric vector starting from 1 providing cell-type labels.}

\item{spike.conc}{spike.conc A vector of theoretical count for each spike-in in one cell (ONLY if spikes = \code{TRUE}).}

\item{CE.range}{A two-element vector of the lower limit and upper limit for the estimated
capture efficiencies (ONLY if spikes = \code{FALSE}, default [0.02, 0.10]).}

\item{use.spikes}{If \code{TRUE}, use spike-ins to estimate capture efficiencies.}

\item{normalize}{Normalization method, either 'ML' (maximum likelihood) or 'TMM' (Robinson et al., 2010).}

\item{GQ.approx}{If \code{TRUE}, use GQ approximation to speed up E-step.}

\item{parallel}{If \code{TRUE}, run in parallel.}

\item{spike}{Observed count matrix for spike-ins, rows represent spike-in molecules, columns represent cells
(ONLY if spikes = \code{TRUE}).}
}
\value{
A list of DE model estimates
}
\description{
Fit the DECENT model with DE assumption
}
\examples{

}
