% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_noDE.R
\name{fitNoDE}
\alias{fitNoDE}
\title{Fitting no-DE model with EM algorithm}
\usage{
fitNoDE(data.obs, CE, normalize, GQ.approx, maxit, parallel)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells.}

\item{CE}{A  vector of capture efficiencies.}

\item{normalize}{Normalization method, either 'ML' (maximum likelihood) or 'TMM' (Robinson et al., 2010).}

\item{GQ.approx}{If \code{TRUE}, use GQ approximation to speed up E-step.}

\item{parallel}{If \code{TRUE}, run in parallel.}
}
\value{
A list of no DE model estimates
}
\description{
Fit the DECENT model with no-DE assumption
}
\examples{

}
