% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputed.R
\name{getImputed}
\alias{getImputed}
\title{Get imputed data}
\usage{
getImputed(data.obs, par.est, sf, CE, XW, tau, parallel = T)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{par.est}{LRT estimates, using par.DE by default}

\item{sf}{A vector containing size factors for all cells}

\item{CE}{A  vector containing capture efficiencies for all cells}

\item{XW}{A matrix containing information for X (cell-type) and W (other covariates)}

\item{tau}{cell-specific estimates (intcp,slope) that link Beta-Binomial dispersion parameter to the mean expression.}

\item{parallel}{If \code{TRUE}, run in parallel}
}
\value{
A matrix containing the mean imputed data.
}
\description{
Get mean imputed data by running E-step once without GQ approximation using the LRT estimates.
}
