% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.R
\name{lrTest}
\alias{lrTest}
\title{Likelihood-ratio test}
\usage{
lrTest(data.obs, out, out2, cell.type, parallel)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{out}{Output object from EM for DE model}

\item{out2}{Output object from EM for no-DE model}

\item{cell.type}{A factor or a integer/numeric vector starting from 1 giving cell-type labels}

\item{parallel}{If \code{TRUE}, run in parallel}
}
\value{
A list including statistics, p-values and parameters for likelihood-ratio test.
}
\description{
Likelihood-ratio test
}
