% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.R
\name{lrTest}
\alias{lrTest}
\title{Likelihood-ratio test}
\usage{
lrTest(data.obs, out, out2, cell.type, parallel)
}
\arguments{
\item{data.obs}{Observed count matrix for endogeneous genes, rows represent genes, columns represent cells}

\item{out}{Output of fitDE, it contains EM algorithm estimates for models with DE between cell-types.}

\item{out2}{Output of fitNoDE, it contains EM algorithm estimates for models without DE between cell-types}

\item{cell.type}{A factor or a integer/numeric vector starting from 1 giving cell-type labels}

\item{parallel}{If \code{TRUE}, run in parallel}
}
\value{
A list containing statistics, p-values and parameter estimates for models with and without DE.
}
\description{
Likelihood-ratio test for DE analysis of scRNA-seq data
}
