% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulated data}
\format{An list of five objects:
\describe{
  \item{\code{data.obs}}{
    a numeric matrix with 1,000 genes and 200 cells: observed count matrix for endogeneous genes
  }
  \item{\code{cell.type}}{
    a numeric vector of length 200: cell-type information
  }
  \item{\code{ercc.obs}}{
    a numeric matrix with 50 spike-ins and 200 cells: observed count matrix for spike-in molecules
  }
  \item{\code{ercc.true}}{
    a numeric vector of length 50: true count of spike-ins added into each cell
  }
  \item{\code{DE.gene}}{
    a logical vector of length 1,000: whether the gene is actually DE or not.
  }
}}
\usage{
sim
}
\description{
A simulated dataset for demonstration. 1,000 genes and 200 cells, with 100 cells in each group.
}
\keyword{datasets}
