% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{zinbGrad}
\alias{zinbGrad}
\title{Gradient of log-likelihood function for ZINB model}
\usage{
zinbGrad(p, y, status, sf, ct)
}
\arguments{
\item{p}{A vector of parameter
p[1] = logistic regression intercept for pi0
p[2] = log mu1
p[3] = log DE par for mu
p[4] = log disp parameter}
}
\description{
Not used currently
}
